feat(category): Implement Categorizing for Pieces of Gallery grid based on category click (#37)
